Add staging info for the buildpack
+ extract parts to different files

[#167706394]
