+++ builder/models.go
@@ -1,28 +1,14 @@
 package builder
 
-import (
-	"fmt"
-)
+type Release struct {
+	DefaultProcessTypes ProcessTypes `yaml:"default_process_types"`
+}
 
-const (
-	Unknown        = "Unknown reason"
-	DetectFailMsg  = "NoAppDetectedError"
-	CompileFailMsg = "BuildpackCompileFailed"
-	ReleaseFailMsg = "BuildpackReleaseFailed"
-
-	FullDetectFailMsg      = "None of the buildpacks detected a compatible application"
-	SupplyFailMsg          = "Failed to run all supply scripts"
-	NoSupplyScriptFailMsg  = "Error: one of the buildpacks chosen to supply dependencies does not support multi-buildpack apps"
-	MissingFinalizeWarnMsg = "Warning: the last buildpack is not compatible with multi-buildpack apps and cannot make use of any dependencies supplied by the buildpacks specified before it"
-	FinalizeFailMsg        = "Failed to run finalize script"
-
-	SystemFailCode   = 1
-	DetectFailCode   = 222
-	CompileFailCode  = 223
-	ReleaseFailCode  = 224
-	SupplyFailCode   = 225
-	FinalizeFailCode = 227
-)
+// StagingInfo is used for export/import droplets
+type StagingInfo struct {
+	DetectedBuildpack string `json:"detected_buildpack" yaml:"detected_buildpack"`
+	StartCommand      string `json:"start_command" yaml:"start_command"`
+}
 
 type ProcessTypes map[string]string
 
@@ -59,39 +45,4 @@ func NewStagingResult(procTypes ProcessTypes, lifeMeta LifecycleMetadata) Stagin
 		ProcessTypes:      procTypes,
 		ExecutionMetadata: "",
 	}
-}
-
-type DescriptiveError struct {
-	ExitCode   int
-	InnerError error
-	Message    string
-}
-
-func (e DescriptiveError) Error() string {
-	if e.InnerError == nil {
-		return fmt.Sprintf("%s: exit status %d", e.Message, e.ExitCode)
-	}
-	return fmt.Sprintf("%s: exit status %d - internal error: %s", e.Message, e.ExitCode, e.InnerError.Error())
-}
-
-var DetectFailErr = DescriptiveError{ExitCode: DetectFailCode, Message: FullDetectFailMsg}
-
-func NewCompileFailError(err error) error {
-	return DescriptiveError{Message: CompileFailMsg, ExitCode: CompileFailCode, InnerError: err}
-}
-
-func NewReleaseFailError(err error) error {
-	return DescriptiveError{Message: ReleaseFailMsg, ExitCode: ReleaseFailCode, InnerError: err}
-}
-
-func NewSupplyFailError(err error) error {
-	return DescriptiveError{Message: SupplyFailMsg, ExitCode: SupplyFailCode, InnerError: err}
-}
-
-func NewFinalizeFailError(err error) error {
-	return DescriptiveError{Message: FinalizeFailMsg, ExitCode: FinalizeFailCode, InnerError: err}
-}
-
-func NewNoSupplyScriptFailError(err error) error {
-	return DescriptiveError{Message: NoSupplyScriptFailMsg, ExitCode: SupplyFailCode, InnerError: err}
 }
